Create hero content
A h1 element, p element, button element and div element have all been
added to the newly created hero-content div element. This will allow
me to add content to this element of the webpage.

The div elements have all been brought in by a tab as it is easier to
see them nested within the body tag.