Remove margins from body, h1 and set min-height
The margin property for both the body and the h1 type selectors has
been set to a value of 0. This ensures that the default margins of
these elements will not interfere with the styling of the webpage.

The min-height of the body has also been set to 100vh. This will
prevent the content from not filling the entire page.