Style hero header, improve readability
The font-size property and value have been added to the hero-header
class selector in order to change the font size of the text. This will
create contrast and a more engaging hero.

The color property has been moved from the hero-info-header class
selector and has been declared after a list of class selectors in
order to avoid repetition and making the code larger than it has to
be. This will also improve efficiency as if the colour needs to be
changed, it will apply to all listed classes rather than having to
manually go through the code to find all instances of said colour to
change.