Change font
The font used throughout the webpage has been changed by adding values
for the font-family to the body selector. This means that all children
of the body will inherit this font. Multiple fonts have been listed
incase the browser doesn't support the one wanted, and sans-serif has
been added as a definite backup.