Specify font size and footer text styling
The font size for the hero subtext and anchor elements has been
specified at a value of 18px to make sure that the text is readable
yet not as large as more important elements, such as the headers.

The footer class selector has been added to the list of selectors
concerning the hero subtext and the anchor elements in order to also
apply the same styling to said footer.