Style testimonial text, improve readability
Font-size, font-style and font-weight have been specified for the text
in the testimonial section in order to make the text more appealing to
look at and read.

The color property has been removed from the information header and
instead added to a list of selectors for the information header and
the testimonial classes. This has been done in order to improve the
readability of the code.