Fix font sizing issues
The font-size property was previously applied to the div containing
the h1 element in the information section. This made the font appear
larger than the h1 element in the hero section despite that element
actually being larger. The font-size and font-weight properties have
now been applied to the h1 element in the information section rather
than the div to fix this issue.