Edit button padding, style placeholder
The button padding has been increased from 5px to 7px as when the
screen was resized it looked too thin.

Padding and font-size have been specified for the placeholder div. The
text has also been aligned to the center of the div and the background
colour and text colour have been slightly adjusted.