Style the image divs in the information section
The height and width of the image divs in the information section have
been specified so that the divs actually take up space on the page.
They have also been assigned a border of the same colour as the sign
up section in order to ensure the colour theme of the page is
consistent. The border radius property has also been added in order to
round the corners of these divs in order to make them look less
jarring and out of place.