Style testimonial section
Padding of 80px has been added to the testimonial section in order to
add space between the content and edge of the div. This will prevent
the information from looking squashed together and creates space that
is easier on the eyes.

The font weight of the quote paragraph element contained in the
testimonial div has been changed to lighter in order to create a
nice contrast between the other headings and bold text on the page.