Style the sign up section
The h1 and p elements have been styled in terms of font size and
colour in order to match the styles specified in the hero section,
except the h1 element has a much smalled font size.

Margin and padding values have been added in order to create space
between the div and the end of the page and also between the div
content and end of the div element to prevent the content from looking
squashed. A border radius property has also been added in order to
round the corners of the div and make it appear much more appealing to
the user.