Make body flex container
The body's display property has been changed to flex and the flex
direction has been set to column. This will be useful in ensuring that
the footer stays at the bottom of the page regardless of the screen
dimensions.

The footer has also been assigned the flex shorthand values of 0 1
auto to ensure that it doesn't grow but is able to shrink to fit
smaller dimensions.