Position header elements with flexbox
The display: flex; rule has been added to the ul type selector in
order to make the list items flex items. A gap of 20px has been
assigned to these list items in order to space them out from one
another.

The margin-right: auto; rule has been added to the hero header div in
order to ensure that it stays on the left side of the page and the
links are all pushed as far possible to the right side of the page.

The header class selector has been added to the file and turned into a
flex container. Padding of 20px has been added to this div in order to
ensure that there is space between the content of the div and the edge
of the element.