Position hero elements with flexbox
A class selector for the hero content div has been added to the file.
The display: flex; rule has been added to this selector in order to
make the div a flex container. Padding of 20px has also been added to
this selecotr in order to ensure space between the content of the div
and the edge of the element. A gap of 30px has also been added to
make sure that there is space between the hero info div and the hero
img div to prevent the page from looking cramped. The content has also
been centrally aligned using justify-content.

A class selector for the hero img div has also been added alongisde
the display: flex; rule to make this div a flex container.