Position information elements with flexbox
The information-header class selector has been added to the file
alongside the margin, display and justify-content properties. The
margin has been set to a value of 50px in order to create space
between the content contained in this div and the edge of the element.
The display has been set to a value of flex in order to make this div
a flex container. The justify-content property has been used to centre
this div.

The media-objects class selector has also been added to the file along
with corresponding margin, display, justify-content and gap
properties. The margin shorthand has a value of 75px 0, allowing me to
give this div a top and bottom margin but not a left and right margin.
The display property has been set to flex in order to make this div a
flex container. The justify-content property has once again been used
to centre the div. A gap of 30px has been added to ensure there is
space between each individual media object div to avoid the page
looking squashed and clunky.

The media-object class selector has also been added to the file with
corresponding width, text-align, display, flex-flow and align-items
properties. The width has been set to 200px to prevent the divs
growing past that size and taking up too much space on the page.
The text-align value has been set to center to ensure the text is
centred underneath the image divs. The display value has been set to
flex in order to make all media-object divs flex containers. The
flex-flow property has been assigned the value of column wrap to
specify that the text should appear below the images and to allow any
text that is larger than the specified width of the flex container to
wrap onto a new flex line. The divs have been centred horizontally
using the align-items property with a value of center.

The height and width values for the image class selector have been
changed to 125px as they previously appeared too large on the page.