Position testimonial elements using flexbox
The display: flex; rule has been added to the testimonial class
selector - making this div a flex container. The flex-flow of this
container has been set to column wrap so that each element appears
below the previous and so that the text is able to wrap onto new flex
lines as needed. The items have also been centrally aligned using
align-items.

The width property has been added to the quote class selector and set
to a value of 775px to ensure that the div doesn't grow to fill the
available space on the page.

The align-self property with a value of flex-end has been added to the
author class so that it appears on the right hand side of the
testimonial container.