Position sign up elements using flexbox
The display: flex; rule has been added to the sign-up class selector
and the sign-up-info class selector in order to make these divs flex
containers.

The flex-flow: column wrap; rule has been added to the sign-up-info
class selector in order to ensure the individual flex items appear
below one another and that any text wraps as needed.

A margin of 0 has been added to the sign-up-text class selector in
order to remove any default padding from this element.

The margin-left: auto; and align-self: center; rules have been applied
to the sign-up-button class selector in order to ensure it is pushed
to the right hand side of the div and that it is centred.